# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/blocked-slots:
    get:
      operationId: listBlockedSlots
      tags: [appointments]
      summary: List blocked slots for a clinic
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      responses:
        "200":
          description: List of blocked slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedSlot"
    post:
      operationId: createBlockedSlot
      tags: [appointments]
      summary: Block a specific slot
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockedSlotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedSlot"
  /clinics/{clinicId}/blocked-slots/{id}:
    delete:
      operationId: deleteBlockedSlot
      tags: [appointments]
      summary: Unblock a slot
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clinics/{clinicId}/blocked-days:
    get:
      operationId: listBlockedDays
      tags: [appointments]
      summary: List blocked days for a clinic
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      responses:
        "200":
          description: List of blocked days
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedDay"
    post:
      operationId: createBlockedDay
      tags: [appointments]
      summary: Block a specific day
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockedDayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedDay"
  /clinics/{clinicId}/blocked-days/{id}:
    delete:
      operationId: deleteBlockedDay
      tags: [appointments]
      summary: Unblock a day
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    ClinicIdParam:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    BlockedSlot:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clinicId, date, startTime, endTime, createdAt]
    BlockedSlotInput:
      type: object
      properties:
        doctorId:
          type: ["integer", "null"]
        date:
          type: string
          minLength: 1
        startTime:
          type: string
          minLength: 1
        endTime:
          type: string
          minLength: 1
        reason:
          type: ["string", "null"]
      required: [date, startTime, endTime]
    BlockedDay:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        date:
          type: string
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clinicId, date, createdAt]
    BlockedDayInput:
      type: object
      properties:
        doctorId:
          type: ["integer", "null"]
        date:
          type: string
          minLength: 1
        reason:
          type: ["string", "null"]
      required: [date]